requirements.txt
certifi==2020.11.8
chardet==3.0.4
DateTime==4.3
idna==2.10
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.4
psycopg2-binary==2.8.6
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
SQLAlchemy==1.3.20
tweepy==3.9.0
urllib3==1.26.2
zope.interface==5.2.0
